requirements.txt
# Inventory Optimization Pipeline Requirements

# Airflow
apache-airflow==3.1.0

# Data Processing
pandas>=2.0.0
numpy>=1.24.0
scipy>=1.11.0

# Database
sqlalchemy>=2.0.0
psycopg2-binary>=2.9.0

# Utilities
python-dotenv>=1.0.0
pydantic>=2.0.0

# Monitoring & Logging
structlog>=23.0.0

# Optional: API integrations
requests>=2.31.0

# Optional: Advanced analytics
scikit-learn>=1.3.0

# Development Dependencies (optional)
pytest>=7.4.0
pytest-cov>=4.1.0
black>=23.0.0
flake8>=6.0.0
mypy>=1.5.0